Allow multiple servers

+++ packages/cluster/scripts/run-consul.sh
@@ -254,8 +254,11 @@ EOF
   },
   "acl": {
     "enabled": true,
-    "default_policy": "deny",
-    "enable_token_persistence": true
+    "default_policy": "allow",
+    "enable_token_persistence": true,
+    "tokens": {
+      "initial_management": "$CONSUL_ACL_MASTER_TOKEN"
+    }
   },
   "telemetry": {
     "prometheus_retention_time": "24h",
@@ -358,11 +361,28 @@ function start_consul {
 }
 
 function bootstrap {
-  local readonly consul_token="$1"
-  log_info "Bootstrapping Consul"
-  echo "${consul_token}" > /tmp/consul.token
-  consul acl bootstrap /tmp/consul.token
-  rm /tmp/consul.token
+  log_info "Waiting for Consul to start"
+  while true; do
+    local readonly consul_leader_addr=$(consul info | grep "leader_addr =" | awk -F'=' '{print $2}' | tr -d  ' ')
+    local readonly consul_leader=$(consul info | grep "leader =" | awk -F'=' '{print $2}' | tr -d  ' ')
+    if [[ -n "$consul_leader_addr" ]]; then
+      log_info "Consul leader elected"
+
+      if [[ "$consul_leader" == "true" ]]; then
+        local readonly consul_token="$1"
+        log_info "Bootstrapping Consul"
+        echo "${consul_token}" > /tmp/consul.token
+        consul acl bootstrap /tmp/consul.token
+        rm /tmp/consul.token
+      fi
+
+      break
+    fi
+
+
+    log_info "Waiting for Consul to start"
+    sleep 1
+  done
 }
 
 # Based on: http://unix.stackexchange.com/a/7732/215969
